Arreglo de relaciones

+++ clubNautico/src/main/java/com/example/services/PersonaService.java
@@ -2,11 +2,11 @@ package com.example.services;
 
 import com.example.personas.Persona;
 import com.example.repositories.PersonaRepository;
-import com.example.services.dto.PersonaDTO;
+import com.example.dto.PersonaDTO;
 import org.modelmapper.ModelMapper;
-import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.beans.BeanUtils;
 
 import java.util.List;
 
@@ -14,45 +14,47 @@ import java.util.List;
 public class PersonaService {
 
     @Autowired
-     private PersonaRepository personaRepo;
+    private PersonaRepository personaRepo;
 
-    private final ModelMapper MODELMAPPER = new ModelMapper();
+    private final ModelMapper modelMapper = new ModelMapper();
 
-    public PersonaDTO convertToDto(Persona persona){
-        return MODELMAPPER.map(persona,PersonaDTO.class);
+    public PersonaDTO convertToDto(Persona persona) {
+        return modelMapper.map(persona, PersonaDTO.class);
     }
 
-    public Persona convertToEntity(PersonaDTO personaDTO){
-        return MODELMAPPER.map(personaDTO,Persona.class);
+    public Persona convertToEntity(PersonaDTO personaDTO) {
+        return modelMapper.map(personaDTO, Persona.class);
     }
 
-    public Persona crearPersona(PersonaDTO personaDTO){
+    public Persona crearPersona(PersonaDTO personaDTO) {
         Persona persona = convertToEntity(personaDTO);
-        Persona personaP = this.personaRepo.save(persona);
-        return personaP;
+        return this.personaRepo.save(persona);
     }
 
-    public void borrarPersona(Integer id){
-        personaRepo.deleteById(id);
+    public void borrarPersona(Integer id) {
+        if (personaRepo.existsById(id)) {
+            personaRepo.deleteById(id);
+        } else {
+            throw new RuntimeException("Persona not found with id " + id);
+        }
     }
 
-    public List<Persona> listarPersonas(){
+    public List<Persona> listarPersonas() {
         return this.personaRepo.findAll();
     }
 
-    public Persona getPersonaById(Integer id){
-        return this.personaRepo.findById(id).get();
+    public Persona getPersonaById(Integer id) {
+        return this.personaRepo.findById(id).orElseThrow(() ->
+                new RuntimeException("Persona not found with id " + id));
     }
 
-    public PersonaDTO actualizarPersona(PersonaDTO personaDTO, Integer idPersona){
-        Persona personaUpdate = this.personaRepo.findById(idPersona).orElseThrow(()
-                -> new RuntimeException("Persona not found with id " + idPersona));
+    public PersonaDTO updatePersona(PersonaDTO personaDTO, Integer idPersona) {
+        Persona personaUpdate = this.personaRepo.findById(idPersona).orElseThrow(() ->
+                new RuntimeException("Persona not found with id " + idPersona));
 
         BeanUtils.copyProperties(personaDTO, personaUpdate);
-        Persona actualizaPersona = this.personaRepo.save(personaUpdate);
-
+        this.personaRepo.save(personaUpdate);
         return convertToDto(personaUpdate);
     }
-
-
 }
+
